Rename schemas to sanity/schemaTypes
Refer to [an opinionated guide to sanity studio](https://www.sanity.io/guides/an-opinionated-guide-to-sanity-studio)
