+++ sanity/schemaTypes/heroType.ts
