+++ sanity/schemaTypes/imageGalleryType.ts
