+++ sanity/schemaTypes/promotionType.ts
