+++ sanity/schemaTypes/textWithIllustrationType.ts
